// src/generated/prisma/models/Order.ts
export type OrderWhereInput = {
  AND?: Prisma.OrderWhereInput | Prisma.OrderWhereInput[]
  OR?: Prisma.OrderWhereInput[]
  NOT?: Prisma.OrderWhereInput | Prisma.OrderWhereInput[]
  id?: Prisma.StringFilter<"Order"> | string
  orderNumber?: Prisma.StringFilter<"Order"> | string
  userId?: Prisma.StringNullableFilter<"Order"> | string | null
  customerEmail?: Prisma.StringFilter<"Order"> | string
  customerName?: Prisma.StringNullableFilter<"Order"> | string | null
  customerPhone?: Prisma.StringNullableFilter<"Order"> | string | null
  status?: Prisma.EnumOrderStatusFilter<"Order"> | $Enums.OrderStatus
  currency?: Prisma.StringFilter<"Order"> | string
  subtotalAmount?: Prisma.DecimalFilter<"Order"> | runtime.Decimal | runtime.DecimalJsLike | number | string
  discountAmount?: Prisma.DecimalFilter<"Order"> | runtime.Decimal | runtime.DecimalJsLike | number | string
  shippingAmount?: Prisma.DecimalFilter<"Order"> | runtime.Decimal | runtime.DecimalJsLike | number | string
  totalAmount?: Prisma.DecimalFilter<"Order"> | runtime.Decimal | runtime.DecimalJsLike | number | string
  customerNote?: Prisma.StringNullableFilter<"Order"> | string | null
  placedAt?: Prisma.DateTimeNullableFilter<"Order"> | Date | string | null
  cancelledAt?: Prisma.DateTimeNullableFilter<"Order"> | Date | string | null
  completedAt?: Prisma.DateTimeNullableFilter<"Order"> | Date | string | null
  createdAt?: Prisma.DateTimeFilter<"Order"> | Date | string
  updatedAt?: Prisma.DateTimeFilter<"Order"> | Date | string
  user?: Prisma.XOR<Prisma.UserNullableScalarRelationFilter, Prisma.UserWhereInput> | null
  items?: Prisma.OrderItemListRelationFilter
  addresses?: Prisma.OrderAddressListRelationFilter
  discounts?: Prisma.OrderDiscountListRelationFilter
  payments?: Prisma.PaymentListRelationFilter
  shipments?: Prisma.ShipmentListRelationFilter
  statusLogs?: Prisma.OrderStatusHistoryListRelationFilter
  sourceCart?: Prisma.XOR<Prisma.CartNullableScalarRelationFilter, Prisma.CartWhereInput> | null
  refundRequests?: Prisma.RefundRequestListRelationFilter
}